rename update board to board; move in board fns
