+++ lib/board.ex
@@ -1,4 +1,4 @@
-defmodule UpdateBoard do
+defmodule Board do
   def handle_board_update(result, board_data, player_symbol) do
     {status, position_to_update} = result
 
@@ -6,5 +6,15 @@ defmodule UpdateBoard do
       :ok -> {:ok, Map.replace!(board_data, position_to_update, player_symbol)}
       :error -> result
     end
+  end
+
+  def has_empty_spaces?(board) do
+    board
+    |> Map.values()
+    |> Enum.any?(fn value -> value == " " end)
+  end
+
+  def is_a_empty_space(board_position, board) do
+    Map.get(board, board_position) == " "
   end
 end
